+++ src/main/java/com/calc/app/service/CalculatorService.java
@@ -7,28 +7,28 @@ import org.springframework.stereotype.Service;
 @Service
 public class CalculatorService {
 
-  public double add(double a, double b) {
+  public Long add(Long a, Long b) {
     validate(a, b);
     return a + b;
   }
 
-  public double subtract(double a, double b) {
+  public Long subtract(Long a, Long b) {
     validate(a, b);
     return a - b;
   }
 
-  public double multiply(double a, double b) {
+  public Long multiply(Long a, Long b) {
     validate(a, b);
     return a * b;
   }
 
-  public double divide(double a, double b) {
+  public Long divide(Long a, Long b) {
     validate(a, b);
     if (b == 0) throw new ZeroDivisionError();
     return a / b;
   }
 
-  private void validate(double a, double b) {
+  private void validate(Long a, Long b) {
     if (a < 0 || b < 0) {
       throw new NegativeNumberException("A or B cannot be negative");
     }
